requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.36.0
psycopg2-binary==2.9.9
fastapi==0.103.2
uvicorn==0.23.2
scikit-learn==1.3.1
pandas==2.1.1
